# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * FarmCity API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Avatar {
  id: number;
  playerId: number;
  skinColor: string;
  hairColor: string;
  hairStyle: string;
  shirtColor: string;
  pantsColor: string;
  /** @nullable */
  hatStyle: string | null;
  /** @nullable */
  accessory: string | null;
}

export interface AvatarInput {
  skinColor: string;
  hairColor: string;
  hairStyle: string;
  shirtColor: string;
  pantsColor: string;
  /** @nullable */
  hatStyle?: string | null;
  /** @nullable */
  accessory?: string | null;
}

export interface AvatarOptions {
  skinColors: string[];
  hairColors: string[];
  hairStyles: string[];
  shirtColors: string[];
  pantColors: string[];
  hatStyles: string[];
  accessories: string[];
}

export interface Player {
  id: number;
  username: string;
  createdAt: string;
  isOnline: boolean;
  avatar?: Avatar | null;
}

export interface PlayerSummary {
  id: number;
  username: string;
  avatar?: Avatar | null;
  /** @nullable */
  posX?: number | null;
  /** @nullable */
  posY?: number | null;
}

export interface RegisterInput {
  username: string;
  password: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export interface AuthResponse {
  player: Player;
  token: string;
}

export interface LogoutResult {
  success: boolean;
}

export interface ChatMessage {
  id: number;
  playerId: number;
  username: string;
  message: string;
  createdAt: string;
}

export interface PlazaStatus {
  onlineCount: number;
  totalPlayers: number;
  messagesLast24h: number;
}

export type GetChatMessagesParams = {
limit?: number;
};