# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FarmCity API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new player
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Player registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /players/{id}:
    get:
      operationId: getPlayer
      tags: [players]
      summary: Get player by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Player data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
...
components:
  schemas:
...
    Avatar:
      type: object
      properties:
        id:
          type: integer
        playerId:
          type: integer
        skinColor:
          type: string
        hairColor:
          type: string
        hairStyle:
          type: string
        shirtColor:
          type: string
        pantsColor:
          type: string
        hatStyle:
          type: ["string", "null"]
        accessory:
          type: ["string", "null"]
      required:
        - id
        - playerId
        - skinColor
        - hairColor
        - hairStyle
        - shirtColor
        - pantsColor
        - hatStyle
        - accessory
...
    Player:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        createdAt:
          type: string
          format: date-time
        isOnline:
          type: boolean
        avatar:
          oneOf:
            - $ref: "#/components/schemas/Avatar"
            - type: "null"
      required:
        - id
        - username
        - createdAt
        - isOnline
...
    RegisterInput:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    LoginInput:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    AuthResponse:
      type: object
      properties:
        player:
          $ref: "#/components/schemas/Player"
        token:
          type: string
      required:
        - player
        - token
...