# lib/api-zod/src/generated/types/authResponse.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * FarmCity API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Player } from './player';

export interface AuthResponse {
  player: Player;
  token: string;
}