# lib/api-zod/src/generated/types/avatar.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * FarmCity API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Avatar {
  id: number;
  playerId: number;
  skinColor: string;
  hairColor: string;
  hairStyle: string;
  shirtColor: string;
  pantsColor: string;
  /** @nullable */
  hatStyle: string | null;
  /** @nullable */
  accessory: string | null;
}