# lib/api-zod/src/generated/types/player.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * FarmCity API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Avatar } from './avatar';

export interface Player {
  id: number;
  username: string;
  createdAt: Date;
  isOnline: boolean;
  avatar?: Avatar | null;
}