# lib/api-zod/src/generated/types/registerInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * FarmCity API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterInput {
  username: string;
  password: string;
}